Refatoração de código.
