+++ Agenda/app/src/main/AndroidManifest.xml
@@ -10,7 +10,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
-        android:name=".AgendaApplication">
+        android:name=".AgendaApplication"
+        tools:ignore="AllowBackup">
         <activity
             android:name=".ui.activity.ListaAlunosActivity"
             android:exported="true">
