+++ Agenda/app/src/main/java/br/com/alura/agenda/AgendaApplication.java
@@ -5,6 +5,7 @@ import android.app.Application;
 import br.com.alura.agenda.dao.AlunoDAO;
 import br.com.alura.agenda.model.Aluno;
 
+@SuppressWarnings("WeakerAccess")
 public class AgendaApplication extends Application {
 
     @Override
