+++ Agenda/app/src/main/java/br/com/alura/agenda/ui/activity/ConstantesActivity.java
@@ -1,5 +1,5 @@
 package br.com.alura.agenda.ui.activity;
 
-public interface ConstantesActivity {
+interface ConstantesActivity {
     String CHAVE_ALUNO = "aluno";
 }
