+++ Agenda/app/src/main/java/br/com/alura/agenda/ui/activity/FormularioAlunoActivity.java
@@ -4,8 +4,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
 import android.widget.EditText;
 
 import androidx.annotation.NonNull;
