+++ Agenda/app/src/main/java/br/com/alura/agenda/ui/adapter/ListaAlunosAdapter.java
@@ -15,7 +15,7 @@ import br.com.alura.agenda.model.Aluno;
 
 public class ListaAlunosAdapter extends BaseAdapter {
     private final List<Aluno> alunos = new ArrayList<>();
-    private Context context;
+    private final Context context;
 
     public ListaAlunosAdapter(Context context) {
         this.context = context;
