+++ Agenda/app/src/main/res/layout/activity_formulario_aluno.xml
@@ -5,30 +5,33 @@
 
 <LinearLayout
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
     android:orientation="vertical">
 
 
     <EditText android:id="@+id/activity_formulario_aluno_nome"
         android:layout_width="match_parent"
         android:layout_margin="8dp"
-        android:hint="Nome"
+        android:hint="@string/nome"
         android:inputType="textCapWords"
-        android:layout_height="wrap_content" />
+        android:layout_height="wrap_content"
+        android:importantForAutofill="no" />
 
     <EditText android:id="@+id/activity_formulario_aluno_telefone"
         android:layout_width="match_parent"
         android:layout_margin="8dp"
-        android:hint="Telefone"
+        android:hint="@string/telefone"
         android:inputType="phone"
-        android:layout_height="wrap_content" />
+        android:layout_height="wrap_content"
+        android:importantForAutofill="no" />
 
     <EditText android:id="@+id/activity_formulario_aluno_email"
         android:layout_width="match_parent"
         android:layout_margin="8dp"
-        android:hint="Email"
+        android:hint="@string/email"
         android:inputType="textEmailAddress"
-        android:layout_height="wrap_content" />
+        android:layout_height="wrap_content"
+        android:importantForAutofill="no" />
 
 </LinearLayout>
