+++ Agenda/app/src/main/res/layout/activity_lista_alunos.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -21,7 +19,8 @@
         android:clickable="true"
         android:layout_alignParentRight="true"
         android:layout_marginRight="25dp"
-        android:src="@drawable/ic_action_novo_aluno"/>
+        android:src="@drawable/ic_action_novo_aluno"
+        android:focusable="true" />
