+++ Agenda/app/src/main/res/menu/activity_formulario_aluno_menu.xml
@@ -5,6 +5,6 @@
     <item android:id="@+id/activity_formulario_aluno_menu_salvar"
         app:showAsAction="always"
         android:icon="@drawable/ic_action_salvar_aluno"
-        android:title="Salvar o aluno" />
+        android:title="@string/salvar_o_aluno" />
 
 </menu>
