+++ Agenda/app/src/main/res/menu/activity_lista_alunos_menu.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android">
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item android:id="@+id/activity_lista_alunos_menu_remover"
-        android:title="Remover" />
+        android:title="@string/remover" />
 </menu>
