+++ Agenda/app/src/main/res/values/strings.xml
@@ -1,3 +1,8 @@
 <resources>
     <string name="app_name">Agenda</string>
+    <string name="nome">Nome</string>
+    <string name="telefone">Telefone</string>
+    <string name="email">Email</string>
+    <string name="salvar_o_aluno">Salvar o aluno</string>
+    <string name="remover">Remover</string>
 </resources>
